fix(sf): que el lector compile bajo gcc con -Wsign-compare (REQ-039.2)
`b + 1 < h.pbag.count` con `b` uint16_t: la promocion a int se compara con
size_t y gcc lo trata como error bajo -Werror. clang no lo emite, por eso el
gate local paso y el CI de ubuntu no. Tres sitios en
SoundFontModulatorReader.h, codigo de 2.4 que hasta hoy solo compilaban los
tests; ahora lo incluye SoundFontManager.cpp, produccion con -Werror.

Spec-Id: REQ-039
Stage-Id: REQ-039.2
Session-Id: b48e3fc4-ed94-45d0-94c9-e08130c58727
Agent: claude-code

+++ audio/src/main/cpp/engines/SoundFontModulatorReader.h
@@ -256,7 +256,7 @@ inline std::vector<RegionModulators> readRegionModulators(const void* data, std:
         Range presetGlobalRange;
         std::vector<Modulator> presetGlobalMods;
 
-        for (std::uint16_t b = bagLo; b + 1 < h.pbag.count && b < bagHi; ++b) {
+        for (std::uint16_t b = bagLo; static_cast<std::size_t>(b) + 1 < h.pbag.count && b < bagHi; ++b) {
             const std::uint16_t genLo = riff::rd16(h.pbag.begin + b * kBagSize);
             const std::uint16_t genHi = riff::rd16(h.pbag.begin + (b + 1) * kBagSize);
             const std::uint16_t modLo = riff::rd16(h.pbag.begin + b * kBagSize + 2);
@@ -273,7 +273,7 @@ inline std::vector<RegionModulators> readRegionModulators(const void* data, std:
                     continue;
                 }
                 hadInstrument = true;
-                if (amount + 1 >= h.inst.count) continue;
+                if (static_cast<std::size_t>(amount) + 1 >= h.inst.count) continue;
 
                 const std::uint16_t iBagLo = riff::rd16(h.inst.begin + amount * kInstSize + 20);
                 const std::uint16_t iBagHi =
@@ -282,7 +282,7 @@ inline std::vector<RegionModulators> readRegionModulators(const void* data, std:
                 Range instGlobalRange;
                 std::vector<Modulator> instGlobalMods;
 
-                for (std::uint16_t ib = iBagLo; ib + 1 < h.ibag.count && ib < iBagHi; ++ib) {
+                for (std::uint16_t ib = iBagLo; static_cast<std::size_t>(ib) + 1 < h.ibag.count && ib < iBagHi; ++ib) {
                     const std::uint16_t iGenLo = riff::rd16(h.ibag.begin + ib * kBagSize);
                     const std::uint16_t iGenHi = riff::rd16(h.ibag.begin + (ib + 1) * kBagSize);
                     const std::uint16_t iModLo = riff::rd16(h.ibag.begin + ib * kBagSize + 2);
